REL-4333 Apply patch from Oracle to use system-default SSL version
Patch is taken unchanged from Oracle's perl-IO-Socket-SSL RPM.

From the patch text itself, it can be deducted that IO::Socket::SSL
comes pre-configured to use its own defined TLS versions by default,
but RPM vendor has changed this to use the system-default TLS version.
Patch was always present in this RPM before, so we bring it back
to venvs, because we can't evaluate consequences of this patch missing.